Frontend password resets and associated bits ...
Frontend and CP both use the same /!/auth action routes.

Devs that want to create custom forms in antlers can, and they will now submit to the same /!/auth endpoint as the built in forms.

Added a PasswordReset class that can track the reset form url, and the redirect after resetting. The notification will use the new url from here.

The frontend will set the form url, if one was specified. This will let users point to a custom reset form. If its not specified they'll just get taken to the built in one.

The CP will tell it to redirect back to the CP after resetting - since thats where they started.

Added a RedirectIfAuthenticated middleware that's just like the one in App but goes to the root. We might add more logic later (thinking sites, etc)