structure/page work ...
Pages can be found by uri in a specific site.

The reference to the entry is stored on the page separately to the entry itself.

A page can be designated as the root (of a structure), which will not have a slug. This replaces the more complicated isParent logic.

The Pages class prepends the parent, which happens by default now, rather than the structure modifying its tree.

Fix duplicate url in toArray

Structure passes the route to each of its pages