Work on actions and assets ...
Actions can have fields, which are displayed in a modal alert.
They can be validated.
Their context variables are sent to Javascript and back to the endpoint when submitting. Simple arrays are used now, instead of objects.

Actions without fields just get a generic "are you sure"

Implement the MoveAsset action, lets the user choose the folder, populated by the container in the context.

When finished with bulk actions, everything gets deselected so that if you moved assets around, it doesn't get confused.

Default modal height to auto