Asset folder fieldtype ...
It gets the container from the top level of the store, rather than a sibling value, like it does in v2.

Since this fieldtype only really gets used inside configuration (and thus at the top level), that's probably fine. No need to introduce a way to get sibling values yet.