Hardcoded fields can be lazy loaded ...
Fields get wrapped in a component that will lazily request preloaded meta data, as well as preprocess the values.

If meta values are passed in, it will just spit back what was given without doing anything.

It only does it for "preloadable" fieldtypes, which are defined by whether they have a preload method, or preloadable static property.

Also, add a fallback for site in the relationship fieldtype.