Only assign ids to sets the first time the bard field is created ...
When live preview is opened/closed, the fields are destroyed and recreated.
This would change the IDs assigned to the sets, making them out of sync to what was initially passed into prosemirror.

The ids and metadata is stored in global state (outside of the fieldtype instance) so it's maintained when destroying.

Fixes #433