Allow fieldtypes to update their meta data in vuex. Use it in the Relationship fieldtype ...
Initial step in fixing reactivity issues in live preview (and maybe eventually other instances)

When switching in and out of live preview, the field components are destroyed and recreated. Anything on the instance will be lost.

In the relationship fieldtype, we'll update the item data in the meta object.
When it's recreated, it'll have it since it gets passed into the fieldtype as a prop.

Ref #433