The same objects should be retrieved every time ...
eg. if you did Collection::findByHandle('diary') and Collection::all(), the diary collection would be loaded twice. The objects would be identical, except they are not literally the same object. If you were to do an === comparison, it would fail.

This would cause 404s when visiting an entry's edit page. The route service provider would compare the entry's collection object to the collection object from the route parameter. They'd be different.

Store paths separately, so we can more easily get a path-to-id mapping.
When we get an item, also put it on the store's items property, so we can avoid excessive cache calls.
When iterating over items to get the paths, we cache them there too.

Entry objects store the collection as a string, so that when they're pulled from the cache, old collection instances aren't brought with it.