Eloquent user implementation is back in core, and is now actually the default. Weird, right? ...
Move the conversion of an eloquent model to statamic user conversion into the repository, so existing models don't need to be touched.

Bring the eloquent user stuff that was already once moved into a separate repo, back into the core. Hello again.

Added a command to generate a migration that adds columns to the stock users table, and adds permission tables.

The user repo manager gets the model from the provider config. It already exists so why have a separate thing?