Pass "handle" into fieldtypes instead of "name" ...
Now, "handle" would be the field handle as defined in the blueprint.

"name" is now a computed and would be what you'd expect to be inside a html element's name attribute. In most cases, it's the same as the handle.
But when you get to nested fields, like grid/replicator/bard, you'd get the array format. eg. gridfield[0][nestedfield]

a "namePrefix" prop is added to let nested fields build on the name