Add expects_root when creating structure
Put expects_root in the file even if its not a collection based structure.
Having it there won't hurt anything. When creating through the CP, if it's a navigation based structure, the toggle won't even be shown and the value won't be submitted.

Since the structure gets added to the collection after its created, having this check in there would always be false.