Entry resource and relationship creating tweaks ...
Creating entries from the relationship field was rendering empty items because the store() response didn't have enough detail. Use the entry resource for that.

Add more keys to the resource so it displays properly.

The saved callback needs to account for it now being a resource and nested under another data key.

The redirect can use the edit url.

The entries fieldtype uses the resource for preloading, too.