Replicator (and Bard) collapsed previews are back ...
Also fixed some inconsistencies between bard and replicator.
Also also fix some styling.

Rather than a getReplicatorPreviewText method in v2, we have a replicatorPreview computed that gets watched and emits an event.

There is an "ok" default for fieldtypes that haven't explicitly defined it. We can even improve on that further.