Add default values for any missing blueprint fields ...
Should solve a long standing friction point where a missing field in a loop would inherit the value from the parent context.

The behavior only happens if the field is in the blueprint, so users should make sure it has what they need.

Collection tag test covers this, but since its in the augmentable trait, it should apply all over the place. In the taxonomy tag, etc.