The get method on the query builder supports an array of columns ...
Similar to Laravel's query builder.

In the Stache it doesn't make a difference to how the object is built, since it's only what you use for the where() clauses that matters. Once the objects are retrieved they already have their data.

However, the objects will automatically use the selected keys when using toAugmentedArray().