Update structure when an entry is deleted by hand ...
As part of the stache watcher, allow handling of deleted items.

Replace the collection's structure with a new class. We grab the structureContents and manipulate that because if we were to do $collection->structure() it would try to new up the structure based on its contents (which now has a deleted entry) and would throw a validation error.

A temporary anonymous structure class is used so we can create a tree not tied to a collection (and thus bypassing validation) for the sole purpose of using the $tree->remove() method, to remove the deleted entry reference. The tree class already has a bunch of smarts to do that.