Allow entries to run callbacks after they're saved but before the event is emitted ...
Plop the entry in the right position in the structure using this callback.
The issue was that in the entrysaved event listeners, it wasnt in the tree yet and its url would be wrong, for example.

Also fix that when removing the parent field should put it back at the top level. It was just leaving it where it was.