Clean up user groups wip ...
No longer need to track the users on the group object. Remove originalUsers and resetOriginalUsers.
Removed addUser and removeUser. We have $user->addToGroup() and $user->removeFromGroup().
The users method is now just a getter that uses the query builder.
Save and delete return true instead of itself to be consistent with other things.
Usergroup repository flushes its cache when saving or deleting groups.