Move Slack notification to a separate job ...
`needs` will make it wait until both jobs are done.

`if always()` is required because by default, if any `need`ed jobs fail, this will get skipped.
The exception is if you add an `if`.

The slack step uses the `workflow-conclusion` action's env var to only send the notification if it fails. The job will still technically run, just not send anything.

Hardcoded the slack status to failure because its never going to be anything else.