Revert 610a41ba and bump composer requirement ...
It's not a big deal to have different composer versions in the GUI and CLI. It seems to work fine.

The problem with using the global composer command was that because we weren't telling it which php binary to use, it would use the system default which might be different from what the CLI and browser were using.