More standard form validation ...
Since we're using Blueprints now, it has validation built in. We'll validate in the controller like everywhere else. Move the honeypot into the controller too.

However, let the data() method continue to filter out fields that aren't defined on the blueprint.
Unlike entries, you can't just add random values to a form submission, so requiring it to be fields from the blueprint is ok.

Pass field display names as attributes into the validator. In the CP this won't be used since we use a custom validation translation file with inline messaging. But here on the front-end, we want to make sure the right names are used.