Set the term as the parent (#2307) ...
Just like on ListedEntry.

Without setting the parent, it would default to the taxonomy itself because Taxonomy::termBlueprint would set itself as the parent. This eventually breaks in the Terms fieldtype when it tries to get the locale from the "entry".

Also rename variable for clarity and explain it could be more than an entry.

Closes #2307
Closes #2377
Closes #2452