Reset the site if it's null. Closes #2408 ...
The entry listing component emits a site-changed event when you change the site filter. This also happens immediately when it's mounted. In 2408, they're using a single-site app, which would have no site filter, so it's immediately null. You'd see the same behavior if you manually clear the site filter on a multisite install.

Introduced in 28d6563