Reset factory properties after making the entry ...
When using the realtime facade more than once, it treats it like a singleton and returns the same object. So, the properties hung around. If you tried to use the factory a second time and didnt call one of the methods, it would inherit whatever you used the first time.

EntryFactory::id('1')->create();
EntryFactory::id('2')->origin('1')->create();
EntryFactory::id('3')->create(); // this would have an origin still set to 1